created login & register forms in frontend. next: send data to backend
