+++ frontend/src/components/accounts/Login.jsx
@@ -1,42 +1,25 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState } from 'react'
 import { Button, FormControl, TextField } from '@mui/material'
 
 const Login = () => {
-    // [0] is the value of the controlled TextField component
-    // [1] is a boolean value indicating whether the TextField should display an error text or not
     const [state, setstate] = useState({
-        username: ["", false],
-        password1: ["", false],
-        password2: ["", false]
+        username: "",
+        password: "",
     })
-
-    // helper texts to give feedback to user if their input is invalid
-    const helperTexts = {
-        username: "Username is already taken.",
-        password1: "Passwords do not match.",
-        password2: "Passwords do not match."
-    }
 
     // update state whenever a TextField changes
     const handleChange = event => {
         setstate({
             ...state,
-            [event.target.name]: [event.target.value, false]
+            [event.target.name]: event.target.value
         })
     }
 
     // login user in (send POST data to backend)
     const handleLogin = () => {
-        const password1 = state.password1[0]
-        const password2 = state.password2[0]
-        if (password1 !== password2) {
-            setstate({
-                ...state,
-                password1: [password1, true],
-                password2: [password2, true]
-            })
-            return
-        }
+        const username = state.username
+        const password = state.password
+        console.log(`TODO`)
     }
 
     return (
@@ -47,18 +30,12 @@ const Login = () => {
                     <TextField 
                         key={key}
                         name={key}
-                        value={state[key][0]}
-                        error={state[key][1]}
-                        label={
-                            key === "username" ? "Username"
-                            : key === "password1" ? "Password"
-                            : "Confirm password"
-                        }
+                        value={state[key]}
+                        label={key === "username" ? "Username" : "Password"}
                         type={key[0] === "p" ? "password" : "text"}
                         size='small'
                         onChange={handleChange}
                         sx={{my: 1}}
-                        helperText={state[key][1] ? helperTexts[key] : ""}
                     />
                 ))}
                 <Button type='submit' variant='contained' onClick={handleLogin}>Login</Button>
